Feb 21

+++ ft_player.c
@@ -6,7 +6,7 @@
 /*   By: rbraaksm <rbraaksm@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/02/07 15:47:58 by rbraaksm       #+#    #+#                */
-/*   Updated: 2020/02/20 17:40:11 by rbraaksm      ########   odam.nl         */
+/*   Updated: 2020/02/21 13:08:41 by rbraaksm      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,8 +35,8 @@ void	player(t_vars *v, double move, char d, unsigned int color)
 		v->play_x += x;
 		v->play_y -= y;
 		my_mlx_pixel_put(v, v->play_x, v->play_y, color);
-		v->map->posy = (int)(v->play_y / v->tile_h);
-		v->map->posx = (int)(v->play_x / v->tile_w);
+		v->map->posy = (v->play_y / (double)v->tile_h);
+		v->map->posx = (v->play_x / (double)v->tile_w);
 		ft_view(v, 0, 0xFFE4E1, ' ');
 	}
 }
